+++ web/src/router/index.ts
@@ -40,6 +40,11 @@ const router = createRouter({
       component: () => import('../views/Logs.vue'),
     },
     {
+      path: '/models',
+      name: 'models',
+      component: () => import('../views/Models.vue'),
+    },
+    {
       path: '/plugins',
       name: 'plugins',
       component: () => import('../views/Plugins.vue'),